bot/requirements.txt
pyopenssl
twisted
requests
